compiler/natives/src/syscall/js: support Go 1.12 1.13 1.14 three
