+++ compiler/natives/src/syscall/js/go112_js.go
@@ -1,0 +1,8 @@
+// +build js
+// +build !go1.13
+
+package js
+
+func (v Value) String() string {
+	return v.internal().String()
+}
